--- .ci/eslint-plugin-zotero-translator/lib/index.js
@@ -1,25 +1,0 @@
-/**
- * @fileoverview Checks Zotero translators for errors and recommended style
- * @author Emiliano Heyns
- */
-"use strict";
-
-//------------------------------------------------------------------------------
-// Requirements
-//------------------------------------------------------------------------------
-
-var requireIndex = require("requireindex");
-
-//------------------------------------------------------------------------------
-// Plugin Definition
-//------------------------------------------------------------------------------
-
-
-// import all rules in lib/rules
-module.exports.rules = requireIndex(__dirname + "/rules");
-
-// import processors
-module.exports.processors = {
-	'.js': require(__dirname + '/processors/zotero-translator'),
-};
-
