--- .ci/eslint-plugin-zotero-translator/lib/rules/prefer-index-of.js
@@ -1,24 +1,0 @@
-'use strict';
-
-const fs = require('fs');
-const path = require('path');
-
-// this is a very simplistic rule to find 'unnecesary use of indexOf' until I find a better eslint plugin that does this
-
-module.exports = function(context) {
-	return {
-		Program: function(node) {
-			const filename = path.resolve(context.getFilename());
-
-			let lineno = 0;
-			for (const line of fs.readFileSync(filename, 'utf-8').split('\n')) {
-				lineno += 1;
-
-				if (line.match(/\.indexOf(.*) *(=+ *-1|!=+ *-1|> *-1|>= *0|< *0)/)) {
-					context.report(node, `Unnecessary '.indexOf()', use '.includes()' instead in line ${lineno}`);
-					break;
-				}
-			}
-		}
-	};
-};
