+++ .ci/package.json
@@ -5,9 +5,9 @@
   "main": "index.js",
   "scripts": {
     "errors": "find .. -maxdepth 1 -name '*.js' -print0 | xargs -0 eslint -c eslintrc --quiet",
-    "test": "find .. -maxdepth 1 -name '*.js' -print0 | xargs -0 eslint -c eslintrc",
+    "test": "find .. -maxdepth 1 -name '*.js' -print0 | xargs -0 eslint",
     "test:self": "find . -path ./node_modules -prune -o -name '*.js' -print0 | xargs -0 eslint -c eslintrc-self",
-    "start": "find .. -maxdepth 1 -name '*.js' -print0 | xargs -0 eslint -c eslintrc --max-warnings 0"
+    "start": "find .. -maxdepth 1 -name '*.js' -print0 | xargs -0 eslint --max-warnings 0"
   },
   "author": "kba+pz",
   "license": "CC0-1.0",
